voice-server/cartesia-agent/requirements.txt
cartesia-line>=0.1.0
python-dotenv>=1.0.0
supabase>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
